requirements.txt
Django==2.2
djangorestframework==3.11.1
django-cors-headers==3.5.0
requests
djangorestframework_simplejwt
django-environ
json